+++ ide/MoonshineSharedCore/src/elements/firefox-debug-adapter/src/adapter/util/fs.ts
@@ -12,25 +12,3 @@ export async function isExecutable(path: string): Promise<boolean> {
 		return false;
 	}
 }
-
-/**
- * Sometimes (on Windows) the temporary directories created for debugging can't be deleted immediately
- * after terminating Firefox, so this method keeps retrying for 500ms.
- */
-export async function tryRemoveRepeatedly(dir: string): Promise<void> {
-	for (var i = 0; i < 5; i++) {
-		try {
-			await fs.remove(dir);
-			log.debug(`Removed ${dir}`);
-			return;
-		} catch (err) {
-			if (i < 4) {
-				log.debug(`Attempt to remove ${dir} failed, will retry in 100ms`);
-				await delay(100);
-			} else {
-				log.debug(`Attempt to remove ${dir} failed, giving up`);
-				throw err;
-			}
-		}
-	}
-}